Keep menu routes within maintainable file-size bounds
Constraint: Team file-size limit requires source files at or below 250 lines.\nRejected: Inline minification | Would hide route responsibilities instead of separating them.\nConfidence: high\nScope-risk: narrow\nDirective: Keep menu endpoint route wiring thin; put reusable menu/admin helpers in focused sibling modules.\nTested: bun test src/routes/menu/__tests__ tests/http/menu && bun run build\nNot-tested: full bun test
