+++ src/routes/menu/menu.ts
@@ -10,206 +10,15 @@
  */
 
 import { Elysia, t } from 'elysia';
-import { asc } from 'drizzle-orm';
-import { MenuItemSchema, MenuResponseSchema, ScopeSchema, type MenuItem, type MenuMeta, type Scope } from './model.ts';
-import { getFrontendMenuItems } from '../../menu/index.ts';
+import { MenuItemSchema, ScopeSchema, type MenuItem, type Scope } from './model.ts';
 import { getMenuConfig, getMenuSource, reloadMenuConfig } from '../../menu/config.ts';
 import { listCustomMenuItems } from '../../menu/custom-store.ts';
-import { db, menuItems } from '../../db/index.ts';
+import { buildMenuItems, readApiMenuItemsFromDb } from './menu-items.ts';
 
 export type MenuExtras = {
   items?: MenuItem[];
   disable?: Iterable<string>;
 };
-
-export const API_TO_STUDIO: ReadonlyArray<readonly [string, string]> = [
-  ['/api/supersede', '/superseded'],
-  ['/api/search', '/search'],
-  ['/api/list', '/feed'],
-  ['/api/reflect', '/playground'],
-  ['/api/threads', '/'],
-  ['/api/traces', '/traces'],
-  ['/api/schedule', '/schedule'],
-  ['/api/plugins', '/plugins'],
-  ['/api/graph', '/map'],
-  ['/api/map3d', '/map'],
-  ['/api/map', '/map'],
-  ['/api/context', '/evolution'],
-  ['/api/stats', '/pulse'],
-];
-
-function studioPathFor(apiPath: string): string | null {
-  for (const [prefix, studio] of API_TO_STUDIO) {
-    if (apiPath === prefix || apiPath.startsWith(prefix + '/')) return studio;
-  }
-  return null;
-}
-
-type RouteLike = { method?: string; path: string; hooks?: { detail?: unknown } };
-type HasRoutes = { routes: RouteLike[] };
-
-const GROUP_RANK: Record<MenuItem['group'], number> = { main: 0, tools: 1, admin: 2, hidden: 3 };
-
-/**
- * Pure scan of Elysia route sources into MenuItems (source='api').
- * Used by tests and exported for callers that need pre-DB scanning.
- */
-export function menuItemsFromRoutes(sources: HasRoutes[]): MenuItem[] {
-  const items: MenuItem[] = [];
-  const seen = new Set<string>();
-
-  for (const src of sources) {
-    for (const route of src.routes) {
-      const detail = (route.hooks?.detail ?? {}) as { menu?: MenuMeta };
-      const menu = detail.menu;
-      if (!menu || !menu.group) continue;
-
-      const studio = studioPathFor(route.path);
-      if (!studio) continue;
-
-      const key = `${menu.group}:${studio}`;
-      if (seen.has(key)) continue;
-      seen.add(key);
-
-      const order =
-        typeof menu.order === 'number' && Number.isFinite(menu.order) ? menu.order : 999;
-      const slug = studio.replace(/^\//, '') || 'home';
-      const label = menu.label ?? slug.charAt(0).toUpperCase() + slug.slice(1);
-
-      items.push({ path: studio, label, group: menu.group, order, source: 'api' });
-    }
-  }
-
-  return items;
-}
-
-/**
- * Match a stored glob pattern (e.g. `vector.*`) against a concrete host
- * (e.g. `vector.foo.com`). Only `*` is treated as a wildcard; other regex
- * metacharacters are escaped.
- */
-export function hostMatches(pattern: string | null | undefined, host: string): boolean {
-  if (pattern == null) return true;
-  const escaped = pattern.replace(/[.+?^${}()|[\]\\]/g, '\\$&').replace(/\*/g, '.*');
-  return new RegExp(`^${escaped}$`).test(host);
-}
-
-/**
- * Does the row's scope match the requested scope filter?
- *
- * - No scope filter -> everything passes (backward compat)
- * - scope='main'   -> rows with scope='main' or scope='both'
- * - scope='sub'    -> rows with scope='sub'  or scope='both'
- */
-export function scopeMatches(rowScope: string, filterScope?: Scope): boolean {
-  if (filterScope == null) return true;
-  if (rowScope === 'both') return true;
-  return rowScope === filterScope;
-}
-
-/**
- * Read API-sourced menu items from the `menu_items` DB table.
- * Only enabled rows are returned. Source is always 'api' for studio consumers.
- * When `host` is provided, rows are filtered to those with null `host` (shown
- * everywhere) or a glob pattern matching the supplied host.
- * When `scope` is provided, rows are filtered to that scope or 'both'.
- */
-export function readApiMenuItemsFromDb(host?: string, scope?: Scope): MenuItem[] {
-  const rows = db
-    .select()
-    .from(menuItems)
-    .orderBy(asc(menuItems.position))
-    .all();
-
-  const items: MenuItem[] = [];
-  for (const row of rows) {
-    if (row.enabled === false) continue;
-    if (host !== undefined && !hostMatches(row.host, host)) continue;
-    if (!scopeMatches(row.scope, scope)) continue;
-    const group = (['main', 'tools', 'admin', 'hidden'] as const).includes(
-      row.groupKey as MenuItem['group'],
-    )
-      ? (row.groupKey as MenuItem['group'])
-      : 'hidden';
-    const item: MenuItem = {
-      id: String(row.id),
-      parentId: row.parentId == null ? null : String(row.parentId),
-      path: row.path,
-      label: row.label,
-      group,
-      order: row.position,
-      source: 'api',
-    };
-    if (row.icon) item.icon = row.icon;
-    if (row.access === 'public' || row.access === 'auth') item.access = row.access;
-    if (row.hidden) item.hidden = true;
-    if (row.studio) item.studio = row.studio;
-    if (row.scope !== 'main') item.scope = row.scope as Scope;
-    if (row.query) {
-      try {
-        const parsed = JSON.parse(row.query);
-        if (parsed && typeof parsed === 'object' && !Array.isArray(parsed)) {
-          const q: Record<string, string> = {};
-          for (const [k, v] of Object.entries(parsed)) {
-            if (typeof v === 'string') q[k] = v;
-          }
-          if (Object.keys(q).length > 0) item.query = q;
-        }
-      } catch {}
-    }
-    items.push(item);
-  }
-  return items;
-}
-
-/**
- * Merge pre-resolved API items with frontend pages, gist extras, and custom
- * items into the final /api/menu response. First-seen wins on dedupe; disable
- * filters any path.
- */
-export function buildMenuItems(
-  apiItems: MenuItem[],
-  extras?: MenuExtras,
-  customItems: MenuItem[] = [],
-): MenuItem[] {
-  const items: MenuItem[] = [];
-  const seen = new Set<string>();
-  const disableSet = new Set<string>(extras?.disable ?? []);
-
-  for (const item of apiItems) {
-    const key = `${item.group}:${item.path}`;
-    if (seen.has(key)) continue;
-    seen.add(key);
-    items.push(item);
-  }
-
-  for (const item of getFrontendMenuItems()) {
-    const key = `${item.group}:${item.path}`;
-    if (seen.has(key)) continue;
-    seen.add(key);
-    items.push(item);
-  }
-
-  if (extras?.items) {
-    for (const item of extras.items) {
-      const key = `${item.group}:${item.path}`;
-      if (seen.has(key)) continue;
-      seen.add(key);
-      items.push(item);
-    }
-  }
-
-  for (const item of customItems) {
-    const key = `${item.group}:${item.path}`;
-    if (seen.has(key)) continue;
-    seen.add(key);
-    items.push({ ...item, added: true } as MenuItem);
-  }
-
-  const filtered = disableSet.size ? items.filter((i) => !disableSet.has(i.path)) : items;
-  filtered.sort((a, b) => GROUP_RANK[a.group] - GROUP_RANK[b.group] || a.order - b.order);
-  return filtered;
-}
 
 const MenuSourceSchema = t.Object({
   url: t.Nullable(t.String()),
@@ -278,3 +87,11 @@ export function createMenuEndpoint() {
 }
 
 export { MenuItemSchema };
+export {
+  API_TO_STUDIO,
+  buildMenuItems,
+  hostMatches,
+  menuItemsFromRoutes,
+  readApiMenuItemsFromDb,
+  scopeMatches,
+} from './menu-items.ts';
